2009-12-21 08:21 UTC+0100 Przemyslaw Czerpak (druzus/at/priv.onet.pl)   * harbour/include/hbgtcore.h   * harbour/include/hbapigt.h     * moved declaration of hb_gtSetDefault() from hbgtcore.h to hbapigt.h
  * harbour/include/hbapi.h
  * harbour/src/vm/hvm.c
    + added new function hb_vmSetDefaultGT()

  * harbour/bin/hb-func.sh
  * harbour/utils/hbmk2/hbmk2.prg
    * use hb_vmSetDefaultGT() to set default GT
    * removed extern declarations for functions defined in hbapi.h

  * harbour/src/vm/dynlibhb.c
    + added support for HB_LIBLOAD()/HB_LIBFREE() in DJGPP 2.04 builds

  * harbour/config/global.mk
    * use 8.3 names for harbour dynamic libraries in all DOS builds

  * harbour/config/dos/djgpp.mk
    + added support for dynamic libraries (DXE) in DJGPP builds.
      DXE is sth similar to DLL in Windows and can be used with
      DJGPP 2.04. I had to make some modifications in DXE header
      files to make it working so it will not work in default DJGPP
      installation. I'll document these modifications in the future.

  * harbour/config/dos/watcom.mk
    * enable creating of harbour dynamic libraries only when HB_BUILD_DLL
      is explictly set by user to 'yes'. In DOS DJGPP and OpenWatcom builds
      DLLs support is not functional yet.

  * harbour/config/dyn.mk
    * added an option to create harbour dynamic libraries from static
      ones instead of .o|.obj files. It's necessary for DXE3GEN which
      internally calls LD which has some limited on maximum size of passed
      parameters.
      TODO: make HB_DYN_LIBS definition common for dyn.mk and lib.mk

+++ harbour/config/dos/watcom.mk
@@ -7,6 +7,7 @@
 OBJ_EXT := .obj
 LIB_PREF :=
 LIB_EXT := .lib
+DYN_EXT := .dll
 
 ifeq ($(HB_BUILD_MODE),cpp)
    CC := wpp386
@@ -70,32 +71,36 @@ ifneq ($(HB_LINKING_RTL),)
    endif
 endif
 
-DY := $(LD)
-DFLAGS += OP quiet SYS cwdllr
-DY_OUT :=
-DLIBS := $(foreach lib,$(HB_USER_LIBS),$(lib))
-DLIBS += $(foreach lib,$(LIBS),$(LIB_DIR)/$(lib))
-DLIBS += $(foreach lib,$(SYSLIBS),$(lib))
-DLIBS := $(strip $(DLIBS))
+ifeq ($(HB_BUILD_DLL),yes)
 
-ifneq ($(DLIBS),)
-   comma := ,
-   DLIBS_COMMA := LIB $(subst $(subst x,x, ),$(comma) ,$(DLIBS))
-else
-   DLIBS_COMMA :=
-endif
+   DY := $(LD)
+   DFLAGS += OP quiet SYS cwdllr
+   DY_OUT :=
+   DLIBS := $(foreach lib,$(HB_USER_LIBS),$(lib))
+   DLIBS += $(foreach lib,$(LIBS),$(LIB_DIR)/$(lib))
+   DLIBS += $(foreach lib,$(SYSLIBS),$(lib))
+   DLIBS := $(strip $(DLIBS))
 
-# NOTE: The empty line directly before 'endef' HAVE TO exist!
-define dyn_object
-   @$(ECHO) $(ECHOQUOTE)FILE '$(file)'$(ECHOQUOTE) >> __dyn__.tmp
+   ifneq ($(DLIBS),)
+      comma := ,
+      DLIBS_COMMA := LIB $(subst $(subst x,x, ),$(comma) ,$(DLIBS))
+   else
+      DLIBS_COMMA :=
+   endif
 
-endef
-define create_dynlib
-   $(if $(wildcard __dyn__.tmp),@$(RM) __dyn__.tmp,)
-   $(foreach file,$^,$(dyn_object))
-   $(DY) $(DFLAGS) $(HB_USER_DFLAGS) NAME '$(subst /,$(DIRSEP),$(DYN_DIR)/$@)' OP implib='$(IMP_FILE)' @__dyn__.tmp $(DLIBS_COMMA)
-endef
+   # NOTE: The empty line directly before 'endef' HAVE TO exist!
+   define dyn_object
+      @$(ECHO) $(ECHOQUOTE)FILE '$(file)'$(ECHOQUOTE) >> __dyn__.tmp
 
-DY_RULE = $(create_dynlib)
+   endef
+   define create_dynlib
+      $(if $(wildcard __dyn__.tmp),@$(RM) __dyn__.tmp,)
+      $(foreach file,$^,$(dyn_object))
+      $(DY) $(DFLAGS) $(HB_USER_DFLAGS) NAME '$(subst /,$(DIRSEP),$(DYN_DIR)/$@)' OP implib='$(IMP_FILE)' @__dyn__.tmp $(DLIBS_COMMA)
+   endef
+
+   DY_RULE = $(create_dynlib)
+
+endif # HB_BUILD_DLL
 
 include $(TOP)$(ROOT)config/common/watcom.mk
